MaasView
Maaş view genel tasarımı eklendi. main menu'nün ismi değiştirldi
